requirements.txt
pyTelegramBotAPI==3.7.3
python-telegram-bot==12.8
pymongo	== 3.11.0
dnspython == 2.0.0
openpyxl == 3.0.5
Pillow == 7.2.0